Edit average learning_curve to allow for non integer parameter values

+++ Studies/plot_average_learning_curve.py
@@ -88,29 +88,30 @@ if __name__ == "__main__":
         X = np.arange(0, args.eval_freq * len(mean[i]), args.eval_freq)
         plt.figure()
         plt.errorbar(X, mean[i], std[i], fmt="--o")
-        plt.title('Reward per step ; '+ args.title +' {}'.format(int(x)))
+        plt.title('Reward per step ; '+ args.title +' {}'.format(float(x)))
         plt.xlabel("timesteps")
         plt.ylabel("reward/step")
 
-        plt.savefig(args.directory + "/visualizations/scores_{}.png".format(int(x)))
+        plt.savefig(args.directory + "/visualizations/scores_{}.png".format(float(x)))
         
     plt.figure()
     for i,x in enumerate(xs):
         print(mean[i])
         print(std[i])
         X = np.arange(0, len(mean[i]))*args.eval_freq
-        plt.errorbar(X, mean[i], std[i], fmt="--o", label="{}".format(int(x)))
+        plt.errorbar(X, mean[i], std[i], fmt="--o", label="{}".format(float(x)))
     plt.xlabel("timesteps")
     plt.ylabel("reward/step")
     plt.legend()
     plt.savefig(args.directory + "/visualizations/scores_{}.png".format(args.title))
 
     fig, ax = plt.subplots(1)
+
     def animate(i):
         fig.clear()
         ax = fig.add_subplot(111)
         ax.errorbar(X, mean[i], std[i], fmt="--o")
-        ax.set_title('Reward per step ; '+ args.title +' {}'.format(int(xs[i])))
+        ax.set_title('Reward per step ; '+ args.title +' {}'.format(float(xs[i])))
         ax.set_xlabel("timesteps")
         ax.set_ylabel("reward/step")
     anim = animation.FuncAnimation(fig, animate, interval=200, frames=len(mean))
